+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Drop `kls_` prefix from `KLS_Conf` fields
 - Drop `tot_hiccups` from `KLS_Stats`
 - Fix direct usage of `_Alignof` in `kls_region.h`
+- Less verbose debug logging
 - Move `CODEOWNERS`, `SECURITY.md` to `.github/`
 - Move `format.sh`, `installer-sh`, `.astylerc` to `scripts/`
 - Bump `invil` to `0.2.29`